+++ packages/adhere-website/src/components/ui/popup/index.jsx
@@ -9,7 +9,7 @@ import PlayGroundPage, {
   FunctionPropsSection,
 } from '@/lib/PlaygroundPage';
 
-import PopupInner from './popup';
+import { PopupInner, PopupClosePreInner } from './popup';
 
 export default () => {
   function boxPanelConfig() {
@@ -17,20 +17,17 @@ export default () => {
       {
         id: `p1`,
         name: `打开`,
-        mode: 'code',
-        scope: { React },
-        cardProps: {
-          description: {
-            title: '打开',
-            info: '打开',
-          },
-        },
-        codeText: `
+        config: [
+          {
+            title: 'index.jsx',
+            mode: 'code',
+            scope: { React },
+            codeText: `
   import React from 'react';
   import { Button } from 'antd';
   import { Popup } from '@baifendian/adhere';
 
-  import PopupInner from './popup';
+  import { PopupInner } from './popup';
 
   <Button
     type="primary"
@@ -61,38 +58,17 @@ export default () => {
   >
     Open Popup
   </Button>
-
-  /*-----------popup.jsx---------*/
-
+    `,
+          },
+          {
+            title: 'popup.jsx',
+            mode: 'code',
+            scope: { React },
+            codeText: `
   import React, { useImperativeHandle, useRef, useState } from 'react';
   import { Popup } from '@baifendian/adhere';
 
   import styles from './index.less';
-
-  function openPopup() {
-    const ref = React.createRef();
-
-    const popup = Popup.create({
-      onCreate: () => {},
-      onBeforeShow: () => {},
-      onAfterShow: () => {},
-      onBeforeClose: () => {
-        return new Promise((resolve) => {
-          resolve();
-        });
-      },
-      onAfterClose: () => {
-        Popup.destroy(popup);
-      },
-      onDestroy: () => {},
-      children: <PopupInner ref={ref} />,
-      zIndex: 9999,
-    });
-
-    ref.current.setPopup(popup);
-
-    popup.show();
-  }
 
   const PopupInner = React.forwardRef((props, ref) => {
     const popupRef = useRef();
@@ -111,7 +87,22 @@ export default () => {
         <div className={styles.Fixed}>
           <a
             onClick={() => {
-              openPopup();
+              const tref = React.createRef();
+
+              const popup = Popup.create({
+                onCreate: () => {},
+                onBeforeShow: () => {},
+                onAfterShow: () => {},
+                onBeforeClose: () => Promise.resolve(),
+                onAfterClose: () => Popup.destroy(popup),
+                onDestroy: () => {},
+                children: <PopupInner ref={tref} />,
+                zIndex: 9999,
+              });
+
+              tref.current.setPopup(popup);
+
+              popup.show();
             }}
           >
             OpenNewPopup
@@ -158,33 +149,198 @@ export default () => {
     );
   });
 
-  export default PopupInner;
+  const PopupClosePreInner = React.forwardRef((props, ref) => {
+    const popupRef = useRef();
 
+    const [id, setId] = useState('');
 
-  /*-----------index.less---------*/
+    useImperativeHandle(ref, () => ({
+      setPopup: (popup) => {
+        popupRef.current = popup;
+        setId(popup.getId());
+      },
+    }));
 
+    return (
+      <div className={styles.Wrap}>
+        <div className={styles.Fixed}>
+          <a
+            onClick={() => {
+              const tref = React.createRef();
+
+              const popup = Popup.create({
+                onCreate: () => {},
+                onBeforeShow: () => {},
+                onAfterShow: () => {},
+                onBeforeClose: () => Promise.resolve(),
+                onAfterClose: () => Popup.destroy(popup),
+                onDestroy: () => {},
+                children: <PopupInner ref={tref} />,
+                zIndex: 9999,
+              });
+
+              tref.current.setPopup(popup);
+
+              popup.showClosePrePopup();
+            }}
+          >
+            OpenNewPopup
+          </a>
+          <div>Popup{id} Title</div>
+          <a
+            onClick={() => {
+              Popup.close(popupRef.current);
+            }}
+          >
+            Close
+          </a>
+        </div>
+
+        <div className={styles.Auto}>
+          <div className="block">
+            <p className="">
+              Here comes popup. You can put here anything, even independent view with its own
+              navigation. Also not, that by default popup looks a bit different on iPhone/iPod and
+              iPad, on iPhone it is fullscreen.
+            </p>
+            <p>
+              Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse faucibus mauris
+              leo, eu bibendum neque congue non. Ut leo mauris, eleifend eu commodo a, egestas ac
+              urna. Maecenas in lacus faucibus, viverra ipsum pulvinar, molestie arcu. Etiam lacinia
+              venenatis dignissim. Suspendisse non nisl semper tellus malesuada suscipit eu et eros.
+              Nulla eu enim quis quam elementum vulputate. Mauris ornare consequat nunc viverra
+              pellentesque. Aenean semper eu massa sit amet aliquam. Integer et neque sed libero
+              mollis elementum at vitae ligula. Vestibulum pharetra sed libero sed porttitor.
+              Suspendisse a faucibus lectus.
+            </p>
+            <p>
+              Duis ut mauris sollicitudin, venenatis nisi sed, luctus ligula. Phasellus blandit nisl
+              ut lorem semper pharetra. Nullam tortor nibh, suscipit in consequat vel, feugiat sed
+              quam. Nam risus libero, auctor vel tristique ac, malesuada ut ante. Sed molestie, est in
+              eleifend sagittis, leo tortor ullamcorper erat, at vulputate eros sapien nec libero.
+              Mauris dapibus laoreet nibh quis bibendum. Fusce dolor sem, suscipit in iaculis id,
+              pharetra at urna. Pellentesque tempor congue massa quis faucibus. Vestibulum nunc eros,
+              convallis blandit dui sit amet, gravida adipiscing libero.
+            </p>
+          </div>
+        </div>
+      </div>
+    );
+  });
+
+  export { PopupInner, PopupClosePreInner };
+            `,
+          },
+          {
+            title: 'index.less',
+            mode: 'code',
+            scope: { React },
+            codeText: `
   .Wrap {
     display: flex;
     flex-direction: column;
 
     > .Fixed {
-      flex-shrink: 0;
-      height: 3rem;
       display: flex;
-      justify-content: space-between;
+      flex-shrink: 0;
       align-items: center;
+      justify-content: space-between;
+      height: 3rem;
     }
 
     > .Auto {
-      font-size: 14px;
       flex-grow: 1;
-      min-height: 0;
-      overflow-y: auto;
-      padding: 1rem;
       box-sizing: border-box;
+      min-height: 0;
+      padding: 1rem;
+      overflow-y: auto;
+      font-size: 14px;
       line-height: 21px;
     }
   }
+            `,
+          },
+        ],
+        cardProps: {
+          description: {
+            title: '打开',
+            info: '打开',
+          },
+        },
+        type: 'PlayGroundMulit',
+        renderChildren: () => (
+          <Button
+            type="primary"
+            onClick={() => {
+              const ref = React.createRef();
+
+              const popup = Popup.create({
+                onCreate: () => {},
+                onBeforeShow: () => {},
+                onAfterShow: () => {},
+                onBeforeClose: () => Promise.resolve(),
+                onAfterClose: () => Popup.destroy(popup),
+                onDestroy: () => {},
+                children: <PopupInner ref={ref} />,
+                zIndex: 9999,
+              });
+
+              ref.current.setPopup(popup);
+
+              popup.show();
+            }}
+          >
+            Open Popup
+          </Button>
+        ),
+      },
+      {
+        id: `p2`,
+        name: `打开后删除之前的实例`,
+        mode: 'code',
+        scope: { React },
+        cardProps: {
+          description: {
+            title: '打开后删除之前的实例',
+            info: '打开后删除之前的实例',
+          },
+        },
+        codeText: `
+  import React from 'react';
+  import { Button } from 'antd';
+  import { Popup } from '@baifendian/adhere';
+
+  import { PopupClosePreInner } from './popup';
+
+  <Button
+    type="primary"
+    onClick={() => {
+      const ref = React.createRef();
+
+      const popup = Popup.create({
+        onCreate: () => {},
+        onBeforeShow: () => {},
+        onAfterShow: () => {},
+        onBeforeClose: () => {
+          return new Promise((resolve) => {
+            resolve();
+          });
+        },
+        onAfterClose: () => {
+          Popup.destroy(popup);
+        },
+        onDestroy: () => {},
+        children: <PopupClosePreInner ref={ref} />,
+        zIndex: 9999,
+      });
+
+      ref.current.setPopup(popup);
+
+      popup.showClosePrePopup();
+    }}
+  >
+    Open Popup
+  </Button>
 
     `,
         type: 'PlayGround',
@@ -198,16 +354,10 @@ export default () => {
                 onCreate: () => {},
                 onBeforeShow: () => {},
                 onAfterShow: () => {},
-                onBeforeClose: () => {
-                  return new Promise((resolve) => {
-                    resolve();
-                  });
-                },
-                onAfterClose: () => {
-                  Popup.destroy(popup);
-                },
+                onBeforeClose: () => Promise.resolve(),
+                onAfterClose: () => Popup.destroy(popup),
                 onDestroy: () => {},
-                children: <PopupInner ref={ref} />,
+                children: <PopupClosePreInner ref={ref} />,
                 zIndex: 9999,
               });
 
@@ -334,6 +484,22 @@ export default () => {
                 returnDesc: '',
               },
               {
+                name: 'PopupFactory.showClosePrePopup',
+                desc: '关闭hi前的Popup在打开Popup',
+                modifier: 'static',
+                params: [
+                  {
+                    name: 'popup',
+                    desc: '使用create创建的popup对象',
+                    type: 'Popup',
+                    defaultVal: '',
+                    required: '',
+                  },
+                ],
+                returnType: 'void',
+                returnDesc: '',
+              },
+              {
                 name: 'PopupFactory.close',
                 desc: '关闭一个Popup',
                 modifier: 'static',
@@ -382,6 +548,22 @@ export default () => {
               {
                 name: 'Popup.show',
                 desc: '打开Popup',
+                modifier: 'public',
+                params: [
+                  {
+                    name: 'popup',
+                    desc: '使用create创建的popup对象',
+                    type: 'Popup',
+                    defaultVal: '',
+                    required: '',
+                  },
+                ],
+                returnType: 'void',
+                returnDesc: '',
+              },
+              {
+                name: 'Popup.showClosePrePopup',
+                desc: '关闭hi前的Popup在打开Popup',
                 modifier: 'public',
                 params: [
                   {