+++ packages/adhere/es/writing-board/index.d.ts
@@ -1,0 +1,2 @@
+export default Model;
+import Model from "@baifendian/adhere-ui-writingboard/es";